Implement CSR read from `barrier` (#5)
CSR reads from `barrier` are used as barriers within a snitch cluster: A core reaching the instruction halts until all other cores have reached the instruction. This PR implements this semantic making it safe to use `-p9` with the snitch runtime. It additionally adds the name of the CSR to the disassembly output.

Some TODOs were left in the code mostly as probably future tasks that will require double-checking in the future. With the way the instruction is used in the snitch runtime, it is already functional.